@ColeMiller1 2024-09-09
> This change is pretty big -- is it possible to break this PR up into smaller chunks?

Sure - there are 3 things. The way `.N` is determined in `isub`, a TODO to make `!logical_i_subset` more efficient, and a somewhat random C change to make `OMP` loop more efficient when the number of threads is set to 1. I can do the two `[.data.table` calls in separate PRs. Is the C change of interest at all?