@MichaelChirico 2024-09-10
> Is the C change of interest at all?

sure! it looks pretty small, should be straightforward to review on its own